Delete obsolete screens and modify existing widgets (intermediary step)
