+++ lib/widgets/features/sign_up/sign_up_method.dart
@@ -94,7 +94,7 @@ class SignupMethodScreen extends StatelessWidget {
                 ),
               ),
               onTap: () {
-                context.push(Routes.signupEmail.path);
+                context.push(Routes.signupCredentials.path);
               },
             ),
           ),
