--- lib/widgets/features/sign_up/sign_up_password.dart
@@ -1,98 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:go_router/go_router.dart';
-import 'package:mm_flutter_app/constants/app_constants.dart';
-import 'package:mm_flutter_app/providers/models/user_registration_model.dart';
-import 'package:mm_flutter_app/widgets/features/sign_up/components/sign_up_template.dart';
-import 'package:mm_flutter_app/widgets/shared/text_form_field_widget.dart';
-import 'package:provider/provider.dart';
-
-class SignupPasswordScreen extends StatefulWidget {
-  const SignupPasswordScreen({super.key});
-
-  @override
-  State<SignupPasswordScreen> createState() => _SignupPasswordScreenState();
-}
-
-class _SignupPasswordScreenState extends State<SignupPasswordScreen> {
-  final GlobalKey<FormState> _formKey = GlobalKey<FormState>();
-  final TextEditingController _pass = TextEditingController();
-  final TextEditingController _confirmPass = TextEditingController();
-  late final UserRegistrationModel _registrationModel;
-  String? _password;
-  String? _confirmPassword;
-
-  @override
-  void initState() {
-    super.initState();
-    _registrationModel = Provider.of<UserRegistrationModel>(
-      context,
-      listen: false,
-    );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final AppLocalizations l10n = AppLocalizations.of(context)!;
-    return SignUpTemplate(
-      progress: SignUpProgress.one,
-      title: l10n.signupPasswordTitle,
-      body: Form(
-        key: _formKey,
-        child: Column(
-          children: [
-            TextFormFieldWidget(
-              isPassword: true,
-              textController: _pass,
-              validator: (val) {
-                if (val == null) {
-                  return 'Empty';
-                }
-                if (val != _confirmPass.text) {
-                  return 'Not Match';
-                }
-                return null;
-              },
-              label: l10n.signupPasswordInputLabel,
-              onChanged: (value) {
-                setState(() {
-                  _password = value;
-                });
-              },
-            ),
-            const SizedBox(
-              height: Insets.paddingSmall,
-            ),
-            TextFormFieldWidget(
-              isPassword: true,
-              textController: _confirmPass,
-              label: l10n.signupPasswordConfirmInputLabel,
-              onChanged: (value) {
-                setState(() {
-                  _confirmPassword = value;
-                });
-              },
-              validator: (val) {
-                if (val == null) {
-                  return 'Empty';
-                }
-                if (val != _pass.text) {
-                  return 'Not Match';
-                }
-                return null;
-              },
-            ),
-          ],
-        ),
-      ),
-      isNextEnabled: (_password?.isNotEmpty ?? false) &&
-          (_confirmPassword?.isNotEmpty ?? false),
-      onNextPressed: () {
-        if (_formKey.currentState!.validate()) {
-          _registrationModel.signUpUserInput.password = _password;
-          context.push(Routes.signupVerification.path);
-        }
-      },
-    );
-  }
-}
